0.5.0 RC1
