+++ backend/src/resource/asset/snippets/blog-frontend-0b390d584a933dd3/asset/editor.js
@@ -7,9 +7,6 @@ export function getContent() {
 }
 
 let allTagsBox = null;
-export function initAllTagsBox() {
-    // allTagsBox = document.getElementById('tags');
-}
 
 export function inputTag(event) {
     if (event.keyCode !== 13)
