+++ backend/src/resource/page/index.html
@@ -17,7 +17,7 @@
     
     <link rel="stylesheet" href="/asset/index-ec4ef66164c33b9c.css">
 
-<link rel="preload" href="/asset/index-8400e80c80322da0_bg.wasm" as="fetch" type="application/wasm" crossorigin="">
-<link rel="modulepreload" href="/asset/index-8400e80c80322da0.js"></head>
+<link rel="preload" href="/asset/index-d3a025b611e964d_bg.wasm" as="fetch" type="application/wasm" crossorigin="">
+<link rel="modulepreload" href="/asset/index-d3a025b611e964d.js"></head>
 <body>
-<script type="module">import init from '/asset/index-8400e80c80322da0.js';init('/asset/index-8400e80c80322da0_bg.wasm');</script></body></html>
+<script type="module">import init from '/asset/index-d3a025b611e964d.js';init('/asset/index-d3a025b611e964d_bg.wasm');</script></body></html>
